+++ FinalExam/src/main/java/com.example.finalexam/service/TeamService.java
@@ -1,7 +1,11 @@
 package com.example.finalexam.service;
 
+import com.example.finalexam.constants.ErrorMessages;
+import com.example.finalexam.exceptions.EntityAlreadyExistsException;
 import com.example.finalexam.exceptions.EntityNotFoundException;
+import com.example.finalexam.exceptions.NoChangesMadeException;
 import com.example.finalexam.model.Team;
+import com.example.finalexam.repository.PlayerRepository;
 import com.example.finalexam.repository.TeamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -9,41 +13,53 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class TeamService {
+public class TeamService implements CrudService<Team> {
     @Autowired
     private TeamRepository teamRepository;
+    @Autowired
+    private PlayerRepository playerRepository;
 
     //Get all teams
-    public List<Team> getAllTeams() {
+    @Override
+    public List<Team> getAll() {
         List<Team> teams = teamRepository.findAll();
         //Check if the list is empty
         if (teams.isEmpty()) {
-            throw new EntityNotFoundException("No teams found");
+            throw new EntityNotFoundException(ErrorMessages.TEAMS_NOT_FOUND_MESSAGE);
         }
         return teams;
     }
 
     //Get a team by ID
-    public Team getTeamById(Long id) {
+    @Override
+    public Team getEntityById(Long id) {
         //Check if team is missing and returning entity or exception message
-        return teamRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Team with ID " + id + " not found"));
+        return teamRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(String.format(ErrorMessages.TEAM_NOT_FOUND_MESSAGE, id)));
     }
 
     //Create a team
-    public Team createTeam(Team team) {
+    @Override
+    public Team createEntity(Team team) {
         //Check if the team already exists
         boolean teamExists = teamRepository.existsByNameAndManagerFullNameAndTeamGroup(
                 team.getName(), team.getManagerFullName(), team.getTeamGroup());
         if (teamExists) {
-            throw new EntityNotFoundException("Team already exists");
+            throw new EntityAlreadyExistsException(ErrorMessages.TEAM_ALREADY_EXISTS_MESSAGE);
         }
         return teamRepository.save(team);
     }
 
     //Update a team
-    public Team updateTeam(Long id, Team updatedTeam) {
+    @Override
+    public Team updateEntity(Long id, Team updatedTeam) {
+        //Fetch the existing team from the repository or return exception
         Team existingTeam = teamRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Team with ID " + id + " not found"));
+                .orElseThrow(() -> new EntityNotFoundException(String.format(ErrorMessages.TEAM_NOT_FOUND_MESSAGE, id)));
+
+        //Check if the fields are not updated(no changes)
+        if (hasNoChanges(updatedTeam, existingTeam)) {
+            throw new NoChangesMadeException(ErrorMessages.NO_CHANGES_MADE_MESSAGE);
+        }
         //Update the team's details
         existingTeam.setName(updatedTeam.getName());
         existingTeam.setManagerFullName(updatedTeam.getManagerFullName());
@@ -53,11 +69,23 @@ public class TeamService {
     }
 
     //Delete a team by ID
-    public void deleteTeam(Long id) {
+    @Override
+    public void deleteEntity(Long id) {
         //Check if the team exists
-        Team team = teamRepository
-                .findById(id).orElseThrow(() -> new EntityNotFoundException("Team with ID " + "id" + " not found"));
+        teamRepository
+                .findById(id).orElseThrow(() -> new EntityNotFoundException(String.format(ErrorMessages.TEAM_NOT_FOUND_MESSAGE, id)));
+        //Check if there are players in the team
+        if (playerRepository.existsByTeamId(id)) {
+            throw new IllegalStateException(ErrorMessages.DELETING_WITH_RELATIONS_MESSAGE);
+        }
         //If exists delete it
         teamRepository.deleteById(id);
+    }
+
+    //Validation method
+    private static boolean hasNoChanges(Team updatedTeam, Team existingTeam) {
+        return existingTeam.getName().equals(updatedTeam.getName()) &&
+                existingTeam.getManagerFullName().equals(updatedTeam.getManagerFullName()) &&
+                existingTeam.getTeamGroup().equals(updatedTeam.getTeamGroup());
     }
 }
